Rename sim driver
